feature/secondary-probability: 2ndary probability feature
Added a secondary prediction pass for the processing
of chat messages.  One of the things that we noticed is
that in order to keep the false positives down I had to
tune the confidence threshold fairly high.  However,
we then noticed that we would get probabilities similar
to the following:

```
{
  'call-counter': 0.5425355063489729,
  'wrong-number': 0.11250503785239177,
  'spanish': 0.87748496752022518,
  'dialer-login': 0.10369462265078187,
  'hang-up': 0.16377986562762845
}
```

Where one of the items would have a relatively high
confidence score (spanish) but it would not reach the
confidence threshold that would keep false positives down.
So, we implemented a secondary prediction pass on the
import text if the highest of the scores that was not
quite above the primary threshold was still greater than
a secondary threshold.  This seems to have greatly
improved the prediction accuracy while keeping false
positives down.

Added feature flags for the aforementioned feature.

Updated the Dockerfile to a multi-stage build to slim
down the images.

Added a fully fleshed out build.sh for automating builds.

The unit file is now parameterized so that multiple
instances of the container can be run on the same machine.